+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - `promiseRace`
 - `setFrom`
 - `size` now supports async iterables
+- `some` now supports async iterables
 - `uniq`
 
 ## 0.3.0 - 2024-02-27